# Improve perf on array operations that can be done in-place