This change updates the way array append and index update are performed in the evaluator to optimize the cases where the updates are known to happen "in-place" and change the original array. These are of the form `set arr += other;` and `set arr w/= i <- val;`, which are quite common in user code. This change introduces two new benchmarks to evaluate the perf impact of the changes, where 10k appends/updates are performed in a loop. The benchmarks show significant improvement, about 78% in append and 98% in update:

```
Array append evaluation time:   [555.13 µs 558.17 µs 561.69 µs]
                        change: [-77.935% -77.702% -77.539%] (p = 0.00 < 0.05)
                        Performance has improved.

Array update evaluation time:   [498.40 µs 499.06 µs 499.84 µs]
                        change: [-98.646% -98.632% -98.621%] (p = 0.00 < 0.05)
                        Performance has improved.
```